+++ webapp/src/components/AssetPage/ListingsTable/ListingsTable.module.css
@@ -17,7 +17,7 @@
   border-bottom: 1px solid var(--secondary);
 }
 
-.tableContainer{
+.tableContainer {
   display: flex;
   flex-direction: column;
   border: 2px solid var(--secondary);
@@ -25,7 +25,7 @@
   margin-top: 48px;
 }
 
-.sortByDropdown{
+.sortByDropdown {
   border-radius: 12px;
   margin-right: 20px;
   justify-content: space-between;
@@ -39,29 +39,27 @@
   right: 20px;
 }
 
-.tabStyle{
+.tabStyle {
   padding-left: 20px;
 }
 
-
- .ui.dropdown > .icon {
+.ui.dropdown > .icon {
   color: white !important;
   height: 100%;
   display: flex;
   align-items: center;
 }
 
-@media (max-width: 768px){
-
-  :global(.ui.button.basic){
+@media (max-width: 768px) {
+  :global(.ui.button.basic) {
     text-align: left;
   }
 
-  .filtertabsContainer{
+  .filtertabsContainer {
     flex-direction: column;
   }
 
-  .tabStyle{
+  .tabStyle {
     padding-left: unset;
   }
 
@@ -72,4 +70,3 @@
     margin-bottom: 10px;
   }
 }
-
