+++ webapp/src/components/AssetPage/ListingsTable/ListingsTable.types.ts
@@ -1,0 +1,10 @@
+import { Item } from '@dcl/schemas'
+
+export enum BelowTabs {
+  LISTINGS = 'listings',
+  OWNERS = 'owners'
+}
+
+export type Props = {
+  item: Item
+}
