+++ webapp/src/modules/translation/locales/en.json
@@ -508,6 +508,7 @@
   },
   "owners_table":{
     "owner": "Owner",
+    "issue_number": "Issue Number",
     "issue_number_asc": "Issue Number ↑",
     "issue_number_desc": "Issue Number ↓",
     "owners": "Owners"
