+++ webapp/src/modules/translation/locales/es.json
@@ -504,6 +504,7 @@
   },
   "owners_table":{
     "owner": "Dueño",
+    "issue_number": "Número de emisión",
     "issue_number_asc": "Número de emisión ↑",
     "issue_number_desc": "Número de emisión ↓",
     "owners": "Dueños"
