+++ webapp/src/modules/translation/locales/zh.json
@@ -505,6 +505,7 @@
   },
   "owners_table":{
     "owner": "所有者",
+    "issue_number": "发行数量",
     "issue_number_asc": "发行数量 ↑",
     "issue_number_desc": "发行数量 ↓",
     "owners": "所有者"
